requirements.txt
streamlit==1.33.0
google-generativeai==0.5.0
python-dotenv
gspread 